requirements.txt
hikari[speedups]==2.0.0.dev101
hikari-lightbulb==1.2.1
uvloop>=0.16; sys_platform != 'nt'